# packages/Rehla/Api/openapi/rehla-v1.yaml
openapi: 3.1.0
info:
  title: Rehla Customer API
  version: 1.0.0
  description: Public & Customer REST API for Rehla Platform
servers:
  - url: /api/v1
    description: V1 Production / Local Server
paths:
...
  /travelers:
    get:
      summary: List customer owned travelers
      operationId: listTravelers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of travelers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Traveler'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: Create new traveler profile
      operationId: createTraveler
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Traveler created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Traveler'
        '401':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
...
  /order-submissions:
    post:
      summary: Submit new order with frozen price and idempotency key
      operationId: submitOrder
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Order created
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/ValidationProblem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - code
        - trace_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        trace_id:
          type: string
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    Traveler:
      type: object
      required:
        - id
        - full_name
        - date_of_birth
        - gender
        - passport_number
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        date_of_birth:
          type: string
        gender:
          type: string
        passport_number:
          type: string
        passport_issued_at:
          type: string
        passport_expires_at:
          type: string
...
  responses:
    Problem:
      description: Standard RFC 7807 Problem Details response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationProblem:
      description: RFC 7807 Validation Problem Details response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'